% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args2vars.R
\name{args2vars}
\alias{args2vars}
\title{Set function argument defaults to global variables}
\usage{
args2vars(
  fn = "current",
  remove = FALSE,
  envir = .GlobalEnv,
  reassign = TRUE,
  run_source_all = TRUE,
  set_dots = NULL,
  parent_fn = sys.function(sys.parent()),
  ...
)
}
\arguments{
\item{fn}{The function for which you want to convert its default
argument values into global variables.
Can be one of:
\describe{
\item{"current" or NULL (default)}{Identifies which file is currently active
within your Rstudio IDE. Then searches for any functions that are defined within
that file and runs \link{args2vars} on each of them. In this case,
the listed returned to the user will be nested by each function name.}
\item{File path}{Works the same as "current" except the user can define which
 R source file to search for functions within
 (e.g. \code{"path/to/file.R"}).}
\item{\link[base]{function}}{A literal R function
 (e.g. \code{function(x) x+1}), or \code{utils::packageDate}}
\item{"parent"}{The function within which \link{args2vars} was called.}
}}

\item{remove}{If \code{TRUE}, will remove globals 
assigned by \code{args2vars} (if supplying the same \code{fn} as before).
If the global does not exist, it will be skipped.}

\item{envir}{the \code{\link[base]{environment}} to use.  See \sQuote{Details}.}

\item{reassign}{If a global variable of the same name already exists, 
reassign its value anyway.}

\item{run_source_all}{Source all R scripts first.}

\item{set_dots}{Set the \code{...} variable to the value of \code{set_dots}.}

\item{parent_fn}{The function within which \code{args2vars} was called.
Only used when \code{fn="parent"}.}

\item{...}{
  Arguments passed on to \code{\link[=source_all]{source_all}}
  \describe{
    \item{\code{path}}{Directory containing R files.}
    \item{\code{pattern}}{File name pattern to search for.}
    \item{\code{packages}}{Packages to load.}
  }}
}
\value{
A named list with each argument's default value, or simply a
\link[base]{stop} function declaring "No default".
}
\description{
Useful for development, where you want to quickly initialize default values
for a function while editing its internal code.
}
\details{
Note the lack of parentheses at the end of the function
(to avoid calling it).
}
\examples{
args2vars(fn = utils::packageDate)
}
