% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_file.R
\name{current_file}
\alias{current_file}
\title{Current file}
\usage{
current_file(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[rstudioapi:rstudio-editors]{rstudioapi::getSourceEditorContext}}
  \describe{
    \item{\code{id}}{The ID of a particular document, as retrieved by \code{documentId()}
or similar. Supported in RStudio 2022.06.0 or newer.}
  }}
}
\value{
File path.
}
\description{
Get the path to the currently active file in the RStudio editor.
}
\examples{
if(interactive()) current_file()
}
