% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoptera-package.R
\docType{package}
\name{devoptera-package}
\alias{devoptera}
\alias{devoptera-package}
\title{devoptera: Practical Tools for R Developers}
\description{
echoverse module: Practical tools for R developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RajLabMSSM/devoptera}
  \item Report bugs at \url{https://github.com/RajLabMSSM/devoptera/issues}
}

}
\author{
\strong{Maintainer}: Brian Schilder \email{brian_schilder@alumni.brown.edu} (\href{https://orcid.org/0000-0001-5949-2191}{ORCID})

Authors:
\itemize{
  \item Jack Humphrey \email{Jack.Humphrey@mssm.edu} (\href{https://orcid.org/0000-0002-6274-6620}{ORCID})
  \item Towfique Raj \email{towfique.raj@mssm.edu} (\href{https://orcid.org/0000-0002-9355-5704}{ORCID})
}

}
\keyword{internal}
