% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Get OS}
\usage{
get_os(lower = TRUE)
}
\arguments{
\item{lower}{Make the OS string lowercase.}
}
\value{
OS type as a character string.
}
\description{
Infer the Operating System (OS) of the current machine being used.
}
\examples{
get_os() 
}
