% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_functions.R
\name{list_functions}
\alias{list_functions}
\title{List functions}
\usage{
list_functions(path = current_file())
}
\arguments{
\item{path}{Path to R source file.}
}
\value{
A list of functions.
}
\description{
List all functions defined within an R source file.
Default to using the currently active file.
}
\examples{
if(interactive()) list_functions()
}
