% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hex.R
\name{make_hex}
\alias{make_hex}
\title{Make hex}
\usage{
make_hex(
  pkg = read.dcf(here::here("DESCRIPTION"), fields = "Package")[1],
  subtitle = "echoverse",
  subplot = NULL,
  save_path = here::here("inst", "hex", "hex.png"),
  new_coords = FALSE,
  n_bats = 20,
  gradient = c("#194f68", "#56ffff"),
  p_size = 14,
  p_x = 1,
  p_y = 1.4,
  s_size = 1,
  s_x = 1,
  s_y = 0.8,
  s_height = s_size,
  s_width = s_size,
  h_fill = "#25355c",
  h_color = "#41c6c8",
  spotlight = TRUE,
  l_alpha = 0.3,
  l_width = 10,
  dpi = 300,
  seed = 1234,
  show_plot = TRUE
)
}
\arguments{
\item{pkg}{Package name.}

\item{subtitle}{Hex sticker subtitle.}

\item{subplot}{subplot}

\item{save_path}{Path to save the hex sticker to.}

\item{new_coords}{Generate new random x/y coordinates to position the bats in 
the background along.}

\item{n_bats}{The number of bats you want to plot in the background.}

\item{gradient}{Two colors to use to create a gradient for coloring the bats.}

\item{p_size}{font size for package name}

\item{p_x}{x position for package name}

\item{p_y}{y position for package name}

\item{s_size}{Convenience argument for controlled both 
\code{s_height} and \code{s_width} at the same time.}

\item{s_x}{x position for subplot}

\item{s_y}{y position for subplot}

\item{s_height}{height for subplot}

\item{s_width}{width for subplot}

\item{h_fill}{color to fill hexagon}

\item{h_color}{color for hexagon border}

\item{spotlight}{whether add spotlight}

\item{l_alpha}{maximum alpha for spotlight}

\item{l_width}{width for spotlight}

\item{dpi}{plot resolution}

\item{seed}{Random seed to make point coordinates reproducible 
(only relevant when \code{new_coords=TRUE}).}

\item{show_plot}{Print the plot once created.}
}
\value{
Plot of class \link[hexSticker]{sticker}.
}
\description{
Make a hex sticker for the 
\href{https://github.com/RajLabMSSM/echolocatoR#the-echoverse}{echoverse}.
}
\examples{
hex <- make_hex(pkg="mypackage",
                save_path = tempfile(fileext = "hex.png"))
}
