% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart.R
\name{restart}
\alias{restart}
\title{Restart R}
\usage{
restart(command = deparse(substitute(cat(paste("Welcome back!")))))
}
\arguments{
\item{command}{A command (as a string) to be run after restarting.}
}
\value{
Null
}
\description{
Restart the R session.
Shallow wrapper for \link[rstudioapi]{restartSession}.
}
\examples{
\dontrun{
restart()
}
}
