% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rma.R
\name{rma}
\alias{rma}
\title{Remove all variables}
\usage{
rma(envir = .GlobalEnv, list = ls(pos = envir, envir = envir))
}
\arguments{
\item{envir}{the \code{\link[base]{environment}} to use.  See \sQuote{details}.}

\item{list}{a character vector (or \code{\link[base]{NULL}}) naming objects to be removed.}
}
\value{
Null
}
\description{
Remove all variables from your environment.
}
\examples{
rma()
}
