% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_info.R
\name{session_info}
\alias{session_info}
\title{Session info}
\source{
\href{https://themockup.blog/posts/2022-04-18-session-info/}{
quarto example}
}
\usage{
session_info(collapse = FALSE, title = "details", add_line = TRUE, ...)
}
\arguments{
\item{collapse}{Wrap the session info output in a collapsible HTML element.}

\item{title}{Title of the collapsed element.}

\item{add_line}{Add a divider line after the session info.}

\item{...}{
  Arguments passed on to \code{\link[sessioninfo:session_info]{sessioninfo::session_info}}
  \describe{
    \item{\code{pkgs}}{Which packages to show. It may be:
\itemize{
\item \code{NULL} or \code{"loaded"}: show all loaded packages,
\item \code{"attached"}: show all attached packages,
\item \code{"installed"}: show all installed packages,
\item a character vector of package names. Their (hard) dependencies are
also shown by default, see the \code{dependencies} argument.
}}
    \item{\code{include_base}}{Include base packages in summary? By default this is
false since base packages should always match the R version.}
    \item{\code{info}}{What information to show, it can be \code{"auto"} to choose
automatically, \code{"all"} to show everything, or a character vector
with elements from:
\itemize{
\item \code{"platform"}: show platform information via \code{\link[sessioninfo:platform_info]{platform_info()}},
\item \code{"packages"}: show package information via \code{\link[sessioninfo:package_info]{package_info()}},
\item \code{"python"}: show Python configuration via \code{\link[sessioninfo:python_info]{python_info()}},
\item \code{"external"}: show information about external software, via
\code{\link[sessioninfo:external_info]{external_info()}}.
}}
    \item{\code{dependencies}}{Whether to include the (recursive) dependencies
as well. See the \code{dependencies} argument of \code{\link[utils:install.packages]{utils::install.packages()}}.}
    \item{\code{to_file}}{Whether to print the session information to a file.
If \code{TRUE} the name of the file will be \code{session-info.txt}, but
\code{to_file} may also be a string to specify the file name.}
  }}
}
\value{
Null.
}
\description{
Wrapper for \link[sessioninfo]{session_info}.
}
\examples{
session_info()
}
