% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_chown.R
\name{set_chown}
\alias{set_chown}
\title{Set owner with \code{chown}}
\usage{
set_chown(
  path,
  whoami = "admin",
  sudo = FALSE,
  pass = Sys.getenv("PASS"),
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to file/folder.}

\item{whoami}{Admin username.}

\item{sudo}{Whether to run commands in \code{sudo} mode.}

\item{pass}{Admin password.}

\item{verbose}{Print messages.}
}
\value{
Null
}
\description{
Set owner with \code{chown}
}
\keyword{internal}
