% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_permissions.R
\name{set_permissions}
\alias{set_permissions}
\title{Set permissions}
\usage{
set_permissions(
  path,
  is_folder = FALSE,
  permissions = list(Windows = "Everyone:(OI)(CI)RX", Mac = "u=rx,go=rx", Linux =
    "u=rx,go=rx", default = "u=rx,go=rx"),
  whoami = "admin",
  pass = Sys.getenv("PASS"),
  sudo = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to file/folder.}

\item{is_folder}{Whether \code{path} is a folder. 
If so, permissions will be set recursively for all files/subfolders.}

\item{permissions}{Scope of permissions to set for each OS type. 
"default" is only used when the OS type is not Windows, Mac, or Linux.}

\item{whoami}{Admin username.}

\item{pass}{Admin password.}

\item{sudo}{Whether to run commands in \code{sudo} mode.}

\item{verbose}{Print messages.}
}
\description{
Robust methods for setting file/folder permissions across multiple OS.
}
\examples{
f <- tempfile()
writeLines(text = letters, con = f)
set_permissions(f)
}
