% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_gcc.R
\name{setup_gcc}
\alias{setup_gcc}
\title{Setup gcc}
\usage{
setup_gcc(
  version = "latest",
  link = "/usr/local/bin/gcc",
  sudo = FALSE,
  pass = Sys.getenv("PASS"),
  overwrite = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{version}{gcc version to use (e.g. 7,8,9,10).
If \code{"latest"}, the latest version of gcc 
currently installed on your machine will be used.}

\item{link}{The path or pathname of the link to be created.
    If \code{"."} (or \code{\link[base]{NULL}}), it is inferred from the
    \code{target} argument, if possible.}

\item{sudo}{Whether to run commands in \code{sudo} mode.}

\item{pass}{Admin password.}

\item{overwrite}{If \code{\link[base:logical]{TRUE}}, an existing link file is overwritten,
    otherwise not.}

\item{verbose}{Print messages.}
}
\value{
A data.frame listing all gcc versions, their paths, 
and which one is currently active.
}
\description{
Set up a symlink to override xcode's clang with a valid version of gcc.
If this doesn't work, first try following
 \href{https://mac.r-project.org/tools/}{these instructions}.
}
\examples{
gcc_df <- setup_gcc(overwrite=FALSE) 
}
